refactor: fixes these issues Typo in the import (enviroment instead of environment) (apikey instead of apiKey) URL variable naming (geoCodingUrl instead of geocodingUrl)
