+++ src/enviroments/environment.prod.ts
@@ -1,0 +1,8 @@
+export const enviroment = {
+  production : false,
+  weatherApi: {
+  key: 'c262d66461ee81e8fce6ac38fa3f12d6',
+  baseUrl: 'https://api.openweathermap.org/data/3.0',
+  geocodingUrl: 'https://api.openweathermap.org/geo/1.0/direct'
+  }
+}
